Rename serverid package to serverinfo
