+++ server/src/com/thoughtworks/go/server/service/plugins/processor/serverinfo/ServerInfoRequestProcessor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.server.service.plugins.processor.serverid;
+package com.thoughtworks.go.server.service.plugins.processor.serverinfo;
 
 import com.thoughtworks.go.plugin.api.request.GoApiRequest;
 import com.thoughtworks.go.plugin.api.response.GoApiResponse;
