+++ server/test/unit/com/thoughtworks/go/server/service/plugins/processor/serverinfo/ServerInfoRequestProcessorTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.server.service.plugins.processor.serverid;
+package com.thoughtworks.go.server.service.plugins.processor.serverinfo;
 
 import com.thoughtworks.go.config.ServerConfig;
 import com.thoughtworks.go.plugin.api.GoPluginIdentifier;
